requirements.txt
Django==4.2.7
python-dotenv
mysqlclient==2.2.1
requests==2.32.3
django-background-tasks==1.2.8